gym: move to trx
