+++ src/libtrx/include/libtrx/game/stats/common.h
@@ -1,3 +1,4 @@
 #pragma once
 
+extern void Stats_StartTimer(void);
 extern void Stats_ObserveItemsLoad(void);
