+++ src/tr1/game/stats/common.h
@@ -12,7 +12,6 @@ int32_t Stats_GetSecrets(void);
 void Stats_ComputeFinal(GF_LEVEL_TYPE level_type, FINAL_STATS *stats);
 bool Stats_CheckAllSecretsCollected(GF_LEVEL_TYPE level_type);
 
-void Stats_StartTimer(void);
 void Stats_UpdateTimer(void);
 void Stats_UpdateSecrets(LEVEL_STATS *stats);
