+++ src/tr2/game/overlay.c
@@ -4,7 +4,6 @@
 #include "game/clock.h"
 #include "game/game.h"
 #include "game/game_flow.h"
-#include "game/gym.h"
 #include "game/inventory.h"
 #include "game/music.h"
 #include "game/objects/common.h"
@@ -16,6 +15,7 @@
 #include "global/vars.h"
 
 #include <libtrx/config.h>
+#include <libtrx/game/gym.h>
 #include <libtrx/game/matrix.h>
 #include <libtrx/game/scaler.h>
 #include <libtrx/utils.h>
